half-decent particle system

+++ Lightning2/Rendering/Particle/Particle.cs
@@ -11,7 +11,7 @@ namespace Lightning2
     /// </summary>
     public class Particle
     {
-        public double Lifetime { get; set; }
+        public float Lifetime { get; set; }
         public Vector2 Position { get; set; }
     }
 }
